test(alignscreen): fix tests after refactor
