+++ plugins/lime-plugin-align/src/components/signalSpeech.js
@@ -1,18 +1,6 @@
 import { useState } from 'preact/hooks';
 import { useInterval } from 'react-use';
-
-let synth = window.speechSynthesis;
-let voices = synth.getVoices();
-
-const speech = (text, lang) => {
-	let utterThis = new SpeechSynthesisUtterance(text);
-	utterThis.pitch = 0.9;
-	utterThis.rate = 1.2;
-	utterThis.voice = voices.filter(x => x.lang === lang)[0];
-	synth.cancel();
-	synth.speak(utterThis);
-};
-
+import { speech } from '../speech';
 
 export const SignalSpeech = ({ signal, className, initialMuted=true }) => {
 	const [ lastSpeech, setLastSpeech ] = useState(0);
